Update Women-in-Cyber.md
---
layout       : blocks/working-session
title        : Women in Cyber
type         : workshop
track        :
technology   : IoT
layout       : blocks/working-session
track        :
technology   :
related-to   : 
status       : Done
when-day     : Thu
organizers   :
participants : Neil Barlow,Ade Yoseman Putra,Francois Raynaud,Lisa Raynaud,Petty Meisari
---


Organisations need women to lead projects, including OWASP, where there is room for more female leaders.

Women currently represent only 11 percent of the cyber security workforce worldwide. This statistic is cause for alarm because it’s a key factor in the massive talent shortage that is impacting this crucially important field. It is estimated that, as of now, there are 1 million unfilled cyber security jobs, and that number is growing fast.

This situation poses serious security risks to governments, businesses, and individuals, which is why there have been increasing efforts in recent years to encourage more people to enter the security field.

Without women pursuing careers in cyber security, the industry is missing out on almost half of the population’s talent pool during a period in history when cybercrime represents a very menacing threat.

## Why

The low number of women in cyber security, and ways that we might encourage more women into the field, have received a lot of attention recently. This led me to ponder why I think it is important that more women become cyber security professionals.

## What

I believe that women and others from a broad range of backgrounds can help to re-formulate some of the fundamental principles of cyber security, to answer in a meaningful way questions such as 'What is it that information security enables organisations and people to do?' 'What are the values we hold dear that information security supports?', or ‘What freedoms can we create through cyber security?’

## Who

The target audience for this Working Session is:

 - IT security professionals
 - Devops professionals
 - Developers
 - Managers
 - Social scientists
 - Creative designers
 - Psychologists
 - Philosophers
 - Organisational and learning specialists
 - Educators