Create Securing-Legacy-Applications.md
Line 8: 'You cannot fix the base of a problem on that system. e.g. Modify the code.' could be clearer.  Do you mean 'You cannot fix the base of a problem on that system; you must modify the code', or do you mean 'You cannot fix the base of a problem on that system by modifying the code'? 

Also, I'm not certain that these are actual or projected outcomes from the Working Session, please confirm.